Update kmeans_clustering_app.py
Animate the convergence of the algorithm. Keep number of features at 2 (you can remove the slider bar). Add dashed vertical line on the elbow plot and the silhouette plot where the optimal number of clusters is. Add a slider for increasing the amount of variance in the clusters, right now the clusters are too obvious and non-overlapping. Add some of the other types of clusters from the images here to show that k-means will not work for non-convex clusters.